Fix argument in delete method of Controller. Fix resolver of MySQl table in IntellijIdea

+++ src/main/java/com/yapollo/acountingofservice/controller/Controller.java
@@ -35,7 +35,7 @@ public class Controller {
        }
 
        @DeleteMapping(path="/{id}")
-       public void delete(@PathVariable(value="id") int id){
+       public void delete(@PathVariable(value="id") Long id){
            receiptService.delete(id);
        }
